HelpRequests are assigned with an owner now. After that, owner data is updated with new values based on what the user set on her HelpRequest.
