+++ conf/urls.py
@@ -11,10 +11,11 @@ from django.views.generic import TemplateView
 from rest_framework import routers
 
 from core import views as core_views
+from core import api as core_api
 
 router = routers.DefaultRouter()
-router.register(r'helprequests', core_views.HelpRequestViewSet)
-router.register(r'helprequestsgeo', core_views.HelpRequestGeoViewSet)
+router.register(r'helprequests', core_api.HelpRequestViewSet)
+router.register(r'helprequestsgeo', core_api.HelpRequestGeoViewSet)
 
 urlpatterns = [
     path('admin/', admin.site.urls),
